Update Forge to use new DeliteFileWriter.
Also updates Forge interpreter IO to use the Hadoop Filesystem API
so that it can also read from different filesystems. We also introduce
Dependencies.scala, a tentative attempt at allowing the user to specify
an of order of operations when there are no data dependencies
(this is particularly useful for pure I/O, for example, or for things
like tic/toc).