Add support for higher-kinded type parameters.
We also add support for type classes with multiple type parameters,
using the convention that the last parameter is the type to be
promoted with methods provided by the type class.

These changes were introduced to support OptiML's TrainingSetLike
typeclass, which has multiple parameters that could not be reliably
inferred without using higher-kinded types.