Changes to add an ALU and Link node. Each new node is added as a new tpe, and hence has its own static apply method. However, in order to have greater codegen control at the DSL spec level, the apply method directly generates code instead of implementing an 'allocates' method.
For this to work, the generated code for scala backend must instantiate
an object of the type created. Currently the Scala backend instantiates
anonymous class objects of the types created, which means that the
generated code basically re-creates the IR. This is a temporary solution
because it makes the Scala backend useless. A permanent solution is to
statically add a template for each node implemented as a codegen node in
the apply method, and have the Scala backend instantiate that template
as the codegen rule.